Add more metadata from Goodreads
I like to have more info than just author and title, so I added it into
the goodreads detail function, along with a bunch of helper functions
right below. The new metadata includes the number of pages in the book,
the original date of publication, and average rating on Goodreads.