Fix insertion code
Previously book insertion was a bit buggy: depending on the value of
org-books-and-to-top, it would either insert at the bottom of the
preceding heading instead of at the top of the current one, or else eat
the final :END: of the preceding book's properties drawer.

This bug is my fault, but it went unnoticed because I was using newlines
between headings. Now that I've deleted the newlines, I uncovered the
bug.

Thankfully, the fix wasn't too difficult, and I managed to simplify some
of lepisma's code as well. I tested it in various settings, and I
believe it now works as intended.