Ask to retry after failed url fetch
Sometimes goodreads fails several times in a row, and it's annoying to
have to call the function and paste the URL again. It makes sense to
have automatic retry.

Since the org-books-add-url function is interactive, the call and retry
capability had to be moved into a separate function that allows
recursive calls.