Simplify helm call
The original version of this function had maps over the helm-org result
in two separate places, but taking them out didn't seem to impact
functionality in any way except make the thing faster.

The intent of the original code appears to have been getting markers out
of the helm-org result, to then pass them to the function in order to
add a book at the proper location. However, helm finds the markers
inside the result just fine, so the maps appear to have been redundant.