Refactor org-books-map-entries
The original function was pretty hairy, but this looked like a good
target for a rewrite with a macro. I haven't tested it with a list of
functions for the FUNC argument, but since this is org-books internal, I
feel that it's not that big of a deal.

Since org-books--get-active-books used this function (the only time it's
used), I changed the funcall to the macro by removing the extra lambda.

I also changed a couple of things in the funcall itself:
- org-get-heading will now ignore tags,
- use 1+ function instead of (+ 1).