Improve LibraryThing title and series retrieval
The previous iteration relied only on information from the 'Wiki' in the
table on the page, and only on one slot at that. The new version tries
to get the book title from the following sources, in order:

1. The 'canonical title' in the Wiki.
2. The 'original title' in the Wiki.
3. The header at the front of the page.

This order appears optimal. The header includes unnecessary information
at times (such as the year of publication), however it is always
present, and is used as a last resort.

I've also added code to handle books without a series. Previously, such
books would still have trailing "()" after the title. Now, standalone
books are treated properly.