Fetch additional LibraryThing details via Amazon
All LibraryThing pages that I've tested have a direct Amazon link. This
can be used to get additional details. Right now I'm using it to fetch
the page count and year (some LibraryThing pages don't include the
publication year in the metadata).

However there is a downside to fetching this extra data from Amazon: it
slows down the process, because now two web pages have to be scraped for
each book. I created a user-customizable variable that lets users opt
out of this functionality.