Rehaul Goodreads scraping
Goodreads has been unreliable for me, and part of the reason is that the
HTML of the web pages when I was inspecting them did not seem to match
what I was getting when sending requests from within Emacs.

This is a rewrite of the Goodreads parser, using the actual received
data as examples. This seems to work much better, though I'm not sure
how liable it is to change.